remove requests import
