+++ sdk/core/azure-core/azure/core/_utils.py
@@ -68,3 +68,28 @@ def _convert_to_isoformat(date_time):
 
     deserialized = deserialized.replace(tzinfo=tzinfo)
     return deserialized
+
+def _case_insensitive_dict(*args, **kwargs):
+    """Return a case-insensitive dict from a structure that a dict would have accepted.
+
+    Rational is I don't want to re-implement this, but I don't want
+    to assume "requests" or "aiohttp" are installed either.
+    So I use the one from "requests" or the one from "aiohttp" ("multidict")
+    If one day this library is used in an HTTP context without "requests" nor "aiohttp" installed,
+    we can add "multidict" as a dependency or re-implement our own.
+    """
+    try:
+        from requests.structures import CaseInsensitiveDict
+
+        return CaseInsensitiveDict(*args, **kwargs)
+    except ImportError:
+        pass
+    try:
+        # multidict is installed by aiohttp
+        from multidict import CIMultiDict
+
+        return CIMultiDict(*args, **kwargs)
+    except ImportError:
+        raise ValueError(
+            "Neither 'requests' or 'multidict' are installed and no case-insensitive dict impl have been found"
+        )
